[fix] 게시판 UI 수정
- 서치바 클릭시 키보드 올리기
- 게시판 플로팅 버튼 서치뷰 올라오면 gone 처리
- 게시판 댓글 inputEditText 변경

#18
