+++ georgeois/app/src/main/java/com/example/georgeois/ui/board/BoardMainFragment.kt
@@ -1,6 +1,7 @@
 package com.example.georgeois.ui.board
 
 import android.annotation.SuppressLint
+import android.content.Context
 import android.graphics.Rect
 import android.os.Bundle
 import android.text.Editable
@@ -8,6 +9,7 @@ import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.view.inputmethod.InputMethodManager
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -15,8 +17,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.georgeois.R
 import com.example.georgeois.databinding.FragmentBoardMainBinding
 import com.example.georgeois.databinding.RowBoardMainBinding
-import com.example.georgeois.ui.chat.SpaceItemDecoration
 import com.example.georgeois.ui.main.MainActivity
+import com.example.georgeois.utill.SpaceItemDecoration
 
 class BoardMainFragment : Fragment() {
     lateinit var fragmentBoardMainBinding: FragmentBoardMainBinding
@@ -40,14 +42,33 @@ class BoardMainFragment : Fragment() {
             val spacingInPixels = resources.getDimensionPixelSize(R.dimen.margin_10dp) // 여백 크기를 리소스에서 가져옴
             recyclerViewBoardMainBoard.addItemDecoration(SpaceItemDecoration(spacingInPixels))
 
+            searchBarBoardMain.run {
+                hint = "검색어를 입력해주세요."
+            }
+            searchViewBoardMain.run {
+                hint = "검색어를 입력해주세요."
+            }
 
             searchViewBoardMain
                 .getEditText()
                 .setOnEditorActionListener { v, actionId, event ->
                     searchBarBoardMain.setText(searchViewBoardMain.getText())
                     searchViewBoardMain.hide()
-                    false
+                    val imm = v.context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+                    imm.hideSoftInputFromWindow(v.windowToken, 0)
+                    true
                 }
+
+            searchViewBoardMain.addTransitionListener { searchView, previousState, newState ->
+                if(newState == com.google.android.material.search.SearchView.TransitionState.SHOWING){
+                    floatingActionButtonBoardMainRegister.visibility = View.GONE
+                }
+                // 서치뷰의 백버튼을 눌러 서치뷰가 사라지고 서치바가 보일 때
+                else if(newState == com.google.android.material.search.SearchView.TransitionState.HIDING){
+                    floatingActionButtonBoardMainRegister.visibility = View.VISIBLE
+                }
+            }
+
 
             searchViewBoardMain.getEditText().addTextChangedListener(object : TextWatcher {
                 override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
@@ -60,6 +81,8 @@ class BoardMainFragment : Fragment() {
                     updateSearchResults(s.toString(), itemList, adapter)
                 }
             })
+
+
 
         }
         return fragmentBoardMainBinding.root
@@ -76,10 +99,10 @@ class BoardMainFragment : Fragment() {
 
         inner class BoardMainViewHolder(val rowBoardMainBinding: RowBoardMainBinding) :
             RecyclerView.ViewHolder(rowBoardMainBinding.root) {
-            val boardTitle :TextView
+            private val boardTitle :TextView = rowBoardMainBinding.textViewRowBoardMainTitle
+
             init {
 
-                boardTitle = rowBoardMainBinding.textViewRowBoardMainTitle
                 rowBoardMainBinding.root.setOnClickListener {
                     val bundle = Bundle()
                     bundle.putString("boardTitle", boardTitle.text.toString())
@@ -110,14 +133,4 @@ class BoardMainFragment : Fragment() {
         }
     }
 
-}
-class SpaceItemDecoration(private val bottomSpace: Int) : RecyclerView.ItemDecoration() {
-    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
-        super.getItemOffsets(outRect, view, parent, state)
-
-        // 마지막 아이템이 아닐 경우에만 바텀 여백을 추가
-        if (parent.getChildAdapterPosition(view) != parent.adapter?.itemCount?.minus(1)) {
-            outRect.bottom = bottomSpace
-        }
-    }
 }
