+++ georgeois/app/src/main/java/com/example/georgeois/utill/SpaceItemDecoration.kt
@@ -1,0 +1,16 @@
+package com.example.georgeois.utill
+
+import android.graphics.Rect
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class SpaceItemDecoration(private val bottomSpace: Int) : RecyclerView.ItemDecoration() {
+    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+        super.getItemOffsets(outRect, view, parent, state)
+
+        // 마지막 아이템이 아닐 경우에만 바텀 여백을 추가
+        if (parent.getChildAdapterPosition(view) != parent.adapter?.itemCount?.minus(1)) {
+            outRect.bottom = bottomSpace
+        }
+    }
+}
